+++ htdocs/ovh/admin/ovh_importinvoice.php
@@ -53,11 +53,12 @@ $endpoint = empty($conf->global->OVH_ENDPOINT)?'ovh-eu':$conf->global->OVH_ENDPO
 
 if ($action == 'setvalue' && $user->admin)
 {
-	$idproduct = GETPOST("OVH_IMPORT_SUPPLIER_INVOICE_PRODUCT_ID");
+	$idproduct = (GETPOST("OVH_IMPORT_SUPPLIER_INVOICE_PRODUCT_ID") > 0 ? GETPOST("OVH_IMPORT_SUPPLIER_INVOICE_PRODUCT_ID") : 0);
 
     $result1=dolibarr_set_const($db, "OVH_THIRDPARTY_IMPORT", GETPOST("OVH_THIRDPARTY_IMPORT"), 'chaine', 0, '', $conf->entity);
     $result2=dolibarr_set_const($db, "OVH_IMPORT_SUPPLIER_INVOICE_PRODUCT_ID", $idproduct, 'chaine', 0, '', $conf->entity);
-    if ($result1 >= 0 && $result2 >= 0)
+    $result3=dolibarr_set_const($db, "OVH_DEFAULT_BANK_ACCOUNT", (GETPOST("OVH_DEFAULT_BANK_ACCOUNT") > 0 ? GETPOST("OVH_DEFAULT_BANK_ACCOUNT") : 0), 'chaine', 0, '', $conf->entity);
+    if ($result1 >= 0 && $result2 >= 0 && $result3 >= 0)
     {
         $mesg='<div class="ok">'.$langs->trans("SetupSaved").'</div>';
     }
@@ -154,6 +155,17 @@ else
         print '</td></tr>';
     }
 
+    if ($conf->banque->enabled)
+    {
+    	$var=!$var;
+    	print '<tr '.$bc[$var].'><td>';
+    	print $langs->trans("OvhDefaultBankAccount").'</td><td>';
+    	print $form->select_comptes($conf->global->OVH_DEFAULT_BANK_ACCOUNT, 'OVH_DEFAULT_BANK_ACCOUNT', 0, '', 1);
+    	print '<td>';
+    	//print $langs->trans("KeepEmptyToSaveLinesAsFreeLines");
+    	print '</td></tr>';
+    }
+
     print '</table>';
 
     dol_fiche_end();
@@ -163,172 +175,6 @@ else
 
 print '</form>';
 
-
-/*
-if ($action == 'preimport')
-{
-    $fuser = new User($db);
-    $result=$fuser->fetch('',$logindol);
-    if ($result <= 0)
-    {
-        print "Bad login user to use\n";
-        exit;
-    }
-
-    try {
-        require_once(DOL_DOCUMENT_ROOT.'/core/lib/functions2.lib.php');
-        $params=getSoapParams();
-        ini_set('default_socket_timeout', $params['response_timeout']);
-
-        $soap = new SoapClient($WS_DOL_URL,$params);
-        try {
-            $language="en";
-            $multisession=false;
-
-            //login
-            $session = $soap->login($conf->global->OVHSMS_NICK, $conf->global->OVHSMS_PASS, $language, $multisession);
-            //if ($session) print '<div class="ok">'.$langs->trans("OvhSmsLoginSuccessFull").'</div><br>';
-            if (! $session) print '<div class="error">Error login did not return a session id</div><br>';
-
-            //logout
-            //$soap->logout($session);
-            //  echo "logout successfull\n";
-
-        }
-        catch(Exception $e)
-        {
-            print '<div class="error">';
-            print 'Error '.$e->getMessage().'<br>';
-            print 'If this is an error to connect to OVH host, check your firewall does not block port required to reach OVH manager (for example port 1664).<br>';
-            print '</div>';
-        }
-
-
-        $result = $soap->billingGetAccessByNic($session);
-        echo "billingGetAccessByNic successfull\n";
-        print_r($result); // your code here ...
-        //billingInvoiceList
-        $result = $soap->billingInvoiceList($session);
-        echo "billingInvoiceList successfull\n";
-        foreach ($result as $i=> $r)
-        {
-            $sql="SELECT rowid ";
-            $sql.=' FROM '.MAIN_DB_PREFIX.'facture_fourn as f';
-            $sql.=" WHERE facnumber like '".$r->billnum."'";
-            $resql = $db->query($sql);
-            $num=0;
-            if ($resql)
-            {
-                $num=$db->num_rows($resql);
-            }
-            if ($num ==0)
-            {
-                //facture n'existe pas
-                $db->begin();
-                $result[$i]->info=$soap->billingInvoiceInfo($session, $r->billnum, null, $r->billingCountry); //on recupere les details
-                $r=$result[$i];
-                $facfou = new FactureFournisseur($db);
-
-                $facfou->ref           = $r->billnum;
-                $facfou->socid         = $id_ovh;
-                $facfou->libelle       = "";
-                $facfou->date          = strtotime($r->date);
-                $facfou->date_echeance = strtotime($r->date);
-                $facfou->note_public   = '';
-
-                $facid = $facfou->create($fuser);
-                if ($facid > 0)
-                {
-                    foreach($r->info->details as $d)
-                    {
-                        $label='<strong>ref :'.$d->service.'</strong><br>'.$d->description.'<br
-    > >';
-                        if($d->start)
-                        $label.='<i>du '.date('d/m/Y',strtotime($d->start));
-                        if($d->end)
-                        $label.=' au '.date('d/m/Y',strtotime($d->end));
-                        $amount=$d->baseprice;
-                        $qty=$d->quantity;
-                        $price_base='HT';
-                        $tauxtva=19.6;
-                        $remise_percent=0;
-                        $fk_product=null;
-                        $ret=$facfou->addline($label, $amount, $tauxtva, $qty, $fk_product,
-                        $remise_percent, '', '', '', 0, $price_base);
-                        if ($ret < 0) $nberror++;
-                        if ($nberror)
-                        {
-                            $db->rollback();
-                            echo "ERROR: ".$facfou->error."\n";
-                        }
-                        else
-                        {
-                            $db->commit();
-                        }
-                    }
-                }
-                else
-                {
-                    $db->rollback();
-                    echo "ERROR: ".$facfou->error."\n";
-                }
-            }
-            else
-            {
-                $row=$db->fetch_array($resql);
-                $facid=$row['rowid'];
-                $facfou = new FactureFournisseur($db);
-                echo "fetching fact $facid ...\n";
-                if($facfou->fetch($facid))
-                {
-                    if($facfou->fk_statut == 0)
-                    {
-                        $ref=dol_sanitizeFileName($facfou->ref);
-                        $upload_dir = $conf->fournisseur->facture->dir_output.'/'.get_exdir($facfou->id,2,0,0,$facfou,'invoice_supplier').$ref;
-
-                        if (! is_dir($upload_dir)) dol_mkdir($upload_dir);
-
-                        if (is_dir($upload_dir))
-                        {
-                            $result[$i]->info=$soap->billingInvoiceInfo($session, $r->billnum, null,
-                            $r->billingCountry); //on recupere les details
-                            $r=$result[$i];
-                            $url=$url_pdf."?reference=".$r->billnum."&passwd=".$r->info->password;
-                            $file_name=($upload_dir."/".$facfou->ref_supplier.".pdf");
-                            echo "$url \n";
-                            if(file_exists($file_name))
-                            {
-                                echo "file $file_name exists !!\n";
-                            }
-                            else{
-
-                                file_put_contents($file_name,file_get_contents($url));
-                            }
-                            //print_r($r->info);
-                        }
-                    }
-                    $facfou->set_valid($fuser);
-                }
-                else{
-                    echo "Imposible d'obtenir la facture $facid \n";
-                }
-                //print_r($facfou);
-            }
-        }
-
-
-
-        //logout
-        if (! empty($conf->global->OVH_OLDAPI)) $soap->logout($session);
-        echo "logout successfull\n";
-
-    } catch(SoapFault $fault) {
-        echo $fault;
-    }
-}
-
-dol_fiche_end();
-*/
 
 llxFooter();